Simplify JPA setup #438
JpaEntityRepositoryBase class that can be used to implement entity-backed repositories. Relationships are implemented through the built-in default repositories forwarding requests to the adjacent resource repositories.
