+++ crnk-core/src/main/java/io/crnk/core/repository/foward/strategy/ForwardingStrategyBase.java
@@ -1,6 +1,12 @@
 package io.crnk.core.repository.foward.strategy;
 
+import io.crnk.core.engine.information.resource.ResourceField;
 import io.crnk.core.engine.internal.utils.PreconditionUtil;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.List;
 
 public class ForwardingStrategyBase {
 
@@ -11,4 +17,14 @@ public class ForwardingStrategyBase {
 		this.context = context;
 	}
 
+	protected <D> Collection<D> getOrCreateCollection(Object source, ResourceField field) {
+		Object property = field.getAccessor().getValue(source);
+		if (property == null) {
+			Class<?> propertyClass = field.getType();
+			boolean isList = List.class.isAssignableFrom(propertyClass);
+			property = isList ? new ArrayList() : new HashSet();
+			field.getAccessor().setValue(source, property);
+		}
+		return (Collection<D>) property;
+	}
 }
