+++ crnk-core/src/test/java/io/crnk/core/mock/repository/TaskList.java
@@ -1,0 +1,7 @@
+package io.crnk.core.mock.repository;
+
+import io.crnk.core.mock.models.Task;
+import io.crnk.core.resource.list.DefaultResourceList;
+
+public class TaskList extends DefaultResourceList<Task> {
+}
