+++ crnk-documentation/src/docs/releases/latest/info.html
@@ -10,5 +10,40 @@ http://127.0.0.1:8080/projects?filter={"OR": {"name": "Great Project", "LE": {"i
 
 More information is available in the <a href="stable/documentation/#_nested_filtering">documentation</a>.
 
+<h4>Revised JPA Setup</h4>
 
+The JPA module has a long history with being the first module of Crnk. For this
+reason it has shown a bit its age and other places evolved quite a bit in the meantime.
+This release comes with a simplified, improved setup. Instead of configuring the JpaModule itself,
+the application can specify repositories extending from JpaEntityRepositoryBase:
 
+<pre class="prettyprint">
+@Component
+public class ScheduleRepositoryImpl extends JpaEntityRepositoryBase<ScheduleEntity, Long> {
+
+    public ScheduleRepositoryImpl() {
+        super(ScheduleEntity.class);
+    }
+}
+</pre>
+
+The JPA module will detect those repositories and make them work like in the past with more explicit
+configuration. The new setup has a number of advantages:
+
+<ul>
+    <li>
+        It is a regular repository like any other. In case of Spring Boot you can, for example,
+        inject it anywhere you like as it is a managed bean.
+    </li>
+    <li>
+        It makes use of the forwarding repository behaviors (see @JsonApiRelation) to
+        forward relationship calls to either side. So in case of customizations, only the
+        resource repositories have to be customized, the relationships will follow automatically.
+    </li>
+    <li>
+        Having a repository at hand allows to do any kind of customizations without having
+        to rely on decorators.
+    </li>
+</ul>
+
+The old configuration mechanism is still in place and can continued to be used.
