--- crnk-integration-examples/spring-boot-example/src/main/java/io/crnk/example/springboot/domain/model/ScheduleDto.java
@@ -1,39 +1,0 @@
-package io.crnk.example.springboot.domain.model;
-
-import io.crnk.core.resource.annotations.JsonApiId;
-import io.crnk.core.resource.annotations.JsonApiResource;
-
-@JsonApiResource(type = "scheduleDto")
-public class ScheduleDto {
-
-	@JsonApiId
-	private Long id;
-
-	private String name;
-
-	private String upperName;
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getUpperName() {
-		return upperName;
-	}
-
-	public void setUpperName(String upperName) {
-		this.upperName = upperName;
-	}
-}
