+++ crnk-integration-examples/spring-boot-example/src/main/resources/application.properties
@@ -2,3 +2,5 @@ crnk.path-prefix=/api
 crnk.default-page-limit=10
 crnk.jpa.exposeAll=false
 crnk.security.enabled=false
+
+logging.level.org.hibernate.SQL=DEBUG
