+++ crnk-integration-examples/wildfly-example/src/main/java/io/crnk/example/wildfly/WildflyModuleProducer.java
@@ -1,11 +1,11 @@
 package io.crnk.example.wildfly;
 
 import io.crnk.core.engine.transaction.TransactionRunner;
-import io.crnk.example.wildfly.model.ScheduleEntity;
+import io.crnk.example.wildfly.endpoints.ScheduleRepository;
+import io.crnk.example.wildfly.endpoints.UserRepository;
 import io.crnk.home.HomeModule;
 import io.crnk.jpa.JpaModule;
 import io.crnk.jpa.JpaModuleConfig;
-import io.crnk.jpa.JpaRepositoryConfig;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Produces;
@@ -14,20 +14,31 @@ import javax.persistence.PersistenceContext;
 
 public class WildflyModuleProducer {
 
-	@PersistenceContext
-	private EntityManager em;
+    @PersistenceContext
+    private EntityManager em;
 
-	@Produces
-	@ApplicationScoped
-	public HomeModule produceHomeModule() {
-		return HomeModule.create();
-	}
+    @Produces
+    @ApplicationScoped
+    public HomeModule produceHomeModule() {
+        return HomeModule.create();
+    }
 
-	@Produces
-	@ApplicationScoped
-	public JpaModule produceJpaModule(TransactionRunner transactionRunner) {
-		JpaModuleConfig config = new JpaModuleConfig();
-		config.addRepository(JpaRepositoryConfig.create(ScheduleEntity.class));
-		return JpaModule.createServerModule(config, em, transactionRunner);
-	}
+    @Produces
+    @ApplicationScoped
+    public JpaModule produceJpaModule(TransactionRunner transactionRunner) {
+        JpaModuleConfig config = new JpaModuleConfig();
+        return JpaModule.createServerModule(config, em, transactionRunner);
+    }
+
+    @Produces
+    @ApplicationScoped
+    public ScheduleRepository scheduleRepository() {
+        return new ScheduleRepository();
+    }
+
+    @Produces
+    @ApplicationScoped
+    public UserRepository userRepository() {
+        return new UserRepository();
+    }
 }
