+++ crnk-integration-examples/wildfly-example/src/main/java/io/crnk/example/wildfly/endpoints/ScheduleRepository.java
@@ -1,0 +1,11 @@
+package io.crnk.example.wildfly.endpoints;
+
+import io.crnk.example.wildfly.model.ScheduleEntity;
+import io.crnk.jpa.JpaEntityRepositoryBase;
+
+public class ScheduleRepository extends JpaEntityRepositoryBase<ScheduleEntity, Long> {
+
+    public ScheduleRepository() {
+        super(ScheduleEntity.class);
+    }
+}
