+++ crnk-jpa/build.gradle
@@ -6,12 +6,19 @@ dependencies {
 	compileOnly project(':crnk-client')
 	compileOnly project(':crnk-data:crnk-data-facet')
 	compileOnly group: 'javax', name: 'javaee-api', version: '7.0'
-	compileOnly group: 'com.querydsl', name: 'querydsl-core', version: '4.1.3'
 	compileOnly group: 'com.querydsl', name: 'querydsl-jpa', version: '4.1.3'
-	compileOnly(group: 'org.hibernate.javax.persistence', name: 'hibernate-jpa-2.1-api', version: '1.0.0.Final')
-	compileOnly(group: 'com.querydsl', name: 'querydsl-apt', version: '4.1.3', classifier: 'jpa')
 
 	testAnnotationProcessor(
+			'com.querydsl:querydsl-apt:4.1.3:jpa',
+			"org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final",
+			"javax.annotation:javax.annotation-api:1.3.2",
+	)
+
+	// necessary for intellij having a broken annotation processing support
+	compileOnly 'com.querydsl:querydsl-apt:4.1.3:jpa'
+	compileOnly "org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final"
+	compileOnly "javax.annotation:javax.annotation-api:1.3.2"
+	annotationProcessor(
 			'com.querydsl:querydsl-apt:4.1.3:jpa',
 			"org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final",
 			"javax.annotation:javax.annotation-api:1.3.2",
@@ -24,8 +31,6 @@ dependencies {
 	testCompile project(':crnk-test')
 	testCompile project(':crnk-data:crnk-data-facet')
 	testCompile group: 'javax', name: 'javaee-api', version: '7.0'
-	testCompile group: 'com.querydsl', name: 'querydsl-core', version: '4.1.3'
-	testCompile group: 'com.querydsl', name: 'querydsl-jpa', version: '4.1.3'
 	testCompile group: 'org.hibernate', name: 'hibernate-core', version: '5.2.10.Final'
 	testCompile group: 'org.springframework', name: 'spring-orm', version: '4.3.1.RELEASE'
 	testCompile group: 'org.springframework', name: 'spring-context', version: '4.3.1.RELEASE'
@@ -34,5 +39,6 @@ dependencies {
 	testCompile group: 'com.h2database', name: 'h2', version: '1.4.187'
 	testCompile(group: 'org.hibernate.javax.persistence', name: 'hibernate-jpa-2.1-api', version: '1.0.0.Final')
 	testCompile(group: 'com.querydsl', name: 'querydsl-apt', version: '4.1.3')
+	testCompile group: 'com.querydsl', name: 'querydsl-jpa', version: '4.1.3'
 }
