+++ crnk-jpa/src/test/java/io/crnk/jpa/integration/AttributeNamingIntTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 
 import java.io.Serializable;
 
-public class AttributeNamingEndToEndTest extends AbstractJpaJerseyTest {
+public class AttributeNamingIntTest extends AbstractJpaJerseyTest {
 
 	@Test
 	public void testCanStoreBasicAttributeValues() {
