+++ crnk-jpa/src/test/java/io/crnk/jpa/model/OneToOneTestEntity.java
@@ -1,31 +1,49 @@
 package io.crnk.jpa.model;
 
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.Id;
 import javax.persistence.OneToOne;
 
 @Entity
 public class OneToOneTestEntity {
 
-	@Id
-	private Long id;
+    @Id
+    private Long id;
 
-	@OneToOne
-	private RelatedEntity oneRelatedValue;
+    /**
+     * uni-directional
+     */
+    @OneToOne(fetch = FetchType.LAZY)
+    private RelatedEntity oneRelatedValue;
 
-	public Long getId() {
-		return id;
-	}
+    /**
+     * bi-directional
+     */
+    @OneToOne(fetch = FetchType.LAZY)
+    private OneToOneOppositeEntity oppositeValue;
 
-	public void setId(Long id) {
-		this.id = id;
-	}
+    public Long getId() {
+        return id;
+    }
 
-	public RelatedEntity getOneRelatedValue() {
-		return oneRelatedValue;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-	public void setOneRelatedValue(RelatedEntity oneRelatedValue) {
-		this.oneRelatedValue = oneRelatedValue;
-	}
+    public RelatedEntity getOneRelatedValue() {
+        return oneRelatedValue;
+    }
+
+    public void setOneRelatedValue(RelatedEntity oneRelatedValue) {
+        this.oneRelatedValue = oneRelatedValue;
+    }
+
+    public OneToOneOppositeEntity getOppositeValue() {
+        return oppositeValue;
+    }
+
+    public void setOppositeValue(OneToOneOppositeEntity oppositeValue) {
+        this.oppositeValue = oppositeValue;
+    }
 }
