+++ crnk-jpa/src/test/java/io/crnk/jpa/repository/JpaRepositoryFilterTest.java
@@ -6,6 +6,7 @@ import io.crnk.jpa.JpaEntityRepository;
 import io.crnk.jpa.JpaModule;
 import io.crnk.jpa.JpaRepositoryConfig;
 import io.crnk.jpa.JpaRepositoryFilterBase;
+import io.crnk.jpa.internal.JpaRepositoryUtils;
 import io.crnk.jpa.model.TestEntity;
 import io.crnk.jpa.query.AbstractJpaTest;
 import io.crnk.jpa.query.JpaQuery;
@@ -23,42 +24,46 @@ import javax.persistence.EntityManager;
 @Transactional
 public class JpaRepositoryFilterTest extends AbstractJpaTest {
 
-	private JpaEntityRepository<TestEntity, Long> repo;
+    private JpaEntityRepository<TestEntity, Long> repo;
 
-	private JpaRepositoryFilterBase filter;
+    private JpaRepositoryFilterBase filter;
 
-	@Override
-	@Before
-	public void setup() {
-		super.setup();
-		repo = new JpaEntityRepository<>(module, JpaRepositoryConfig.create(TestEntity.class));
-	}
+    @Override
+    @Before
+    public void setup() {
+        super.setup();
 
-	@SuppressWarnings("unchecked")
-	@Test
-	public void test() {
-		QuerySpec querySpec = new QuerySpec(TestEntity.class);
-		ResourceList<TestEntity> list = repo.findAll(querySpec);
-		Assert.assertEquals(5, list.size());
+        JpaRepositoryConfig<TestEntity> config = JpaRepositoryConfig.create(TestEntity.class);
+        JpaRepositoryUtils.setDefaultConfig(module.getConfig(), config);
+        repo = new JpaEntityRepository<>(config);
+        repo.setResourceRegistry(resourceRegistry);
+    }
 
-		Mockito.verify(filter, Mockito.times(1)).filterQuerySpec(Mockito.eq(repo), Mockito.eq(querySpec));
-		Mockito.verify(filter, Mockito.times(1)).filterResults(Mockito.eq(repo), Mockito.eq(querySpec), Mockito.eq(list));
-		Mockito.verify(filter, Mockito.times(1)).filterExecutor(Mockito.eq(repo), Mockito.eq(querySpec),
-				Mockito.any(JpaQueryExecutor.class));
-		Mockito.verify(filter, Mockito.times(1)).filterTuples(Mockito.eq(repo), Mockito.eq(querySpec), Mockito.anyList());
-		Mockito.verify(filter, Mockito.times(1)).filterQuery(Mockito.eq(repo), Mockito.eq(querySpec),
-				Mockito.any(JpaQuery.class));
-	}
+    @SuppressWarnings("unchecked")
+    @Test
+    public void test() {
+        QuerySpec querySpec = new QuerySpec(TestEntity.class);
+        ResourceList<TestEntity> list = repo.findAll(querySpec);
+        Assert.assertEquals(5, list.size());
 
-	@Override
-	protected void setupModule(JpaModule module) {
-		filter = Mockito.spy(new JpaRepositoryFilterBase());
-		module.addFilter(filter);
-	}
+        Mockito.verify(filter, Mockito.times(1)).filterQuerySpec(Mockito.eq(repo), Mockito.eq(querySpec));
+        Mockito.verify(filter, Mockito.times(1)).filterResults(Mockito.eq(repo), Mockito.eq(querySpec), Mockito.eq(list));
+        Mockito.verify(filter, Mockito.times(1)).filterExecutor(Mockito.eq(repo), Mockito.eq(querySpec),
+                Mockito.any(JpaQueryExecutor.class));
+        Mockito.verify(filter, Mockito.times(1)).filterTuples(Mockito.eq(repo), Mockito.eq(querySpec), Mockito.anyList());
+        Mockito.verify(filter, Mockito.times(1)).filterQuery(Mockito.eq(repo), Mockito.eq(querySpec),
+                Mockito.any(JpaQuery.class));
+    }
 
-	@Override
-	protected JpaQueryFactory createQueryFactory(EntityManager em) {
-		return QuerydslQueryFactory.newInstance();
-	}
+    @Override
+    protected void setupModule(JpaModule module) {
+        filter = Mockito.spy(new JpaRepositoryFilterBase());
+        module.addFilter(filter);
+    }
+
+    @Override
+    protected JpaQueryFactory createQueryFactory(EntityManager em) {
+        return QuerydslQueryFactory.newInstance();
+    }
 
 }
